+++ package.json
@@ -74,23 +74,6 @@
 		"format": "prettier --check \"./**/*.{ts,tsx,js,css,json,md}\"",
 		"format:fix": "prettier --write  \"./**/*.{ts,tsx,js,css,json,md}\""
 	},
-  "jest": {
-    "collectCoverage": true,
-    "testPathIgnorePatterns": [
-      "<rootDir>/tests/fixtures/"
-    ],
-    "coveragePathIgnorePatterns": [
-      "<rootDir>/tests/fixtures/"
-    ],
-    "reporters": [
-      "default",
-      "jest-junit"
-    ],
-    "testEnvironment": "node",
-    "setupFilesAfterEnv": [
-      "./jest.setup.js"
-    ]
-  },
   "description": "Basic working repo structure for Adobe teams to use as a starting point for their own plugins, and available as a github template to the @adobe github org.",
   "directories": {
     "test": "test"
