+++ src/commands/__fixtures__/sample_mesh.json
