--- test/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "rules": {
-      "node/no-unpublished-require": 0
-  }
-}
